refactor(tests): collapse framework-support-matrix drift check into a pure pytest module
The matrix drift check is conceptually a unit test, not a build-time
generator like scripts/generate_pyproject.py. Merge the generator logic
into the existing test module and drop the scripts/ file and the
spec_from_file_location bridge (the only such pattern in the repo).

When the drift test fails, developers regenerate the block between the
BEGIN GENERATED / END GENERATED markers with a coding agent; the doc's
"Where" and "To promote a cell" sections now describe that workflow.
