+++ docs/guides/data-operation-patterns/framework-support-matrix.md
@@ -5,8 +5,8 @@ One-page lookup for "does operation *X* work on framework *Y*?". Rows are the te
 **What**: A capability matrix that mirrors the `supported_*()` class methods declared on every framework test class. If a cell is ✓, the shared reference-based test suite runs that subtype against that framework and compares the result to PyArrow. If a cell is ✗, the framework either cannot express that subtype or has been deliberately excluded (see [Known divergences](known-divergences.md)).
 **When**: Use before picking an op/framework pair, before adding a new framework implementation for an existing op, or while debugging why a feature resolves on one framework but skips on another.
 **Why**: The authoritative information lives in `supported_agg_types()` / `supported_ops()` / `supported_offset_types()` / `supported_rank_types()` overrides scattered across ten operation directories. This page flattens them into a single table and keeps it in sync via a drift check.
-**Where**: The tables below are generated by `scripts/check_framework_support_matrix.py`. Do not edit the block between the `BEGIN GENERATED` and `END GENERATED` markers directly. After changing any `supported_*()` override, run the script from an environment that can import the project package, e.g. `uv run python scripts/check_framework_support_matrix.py` (or `python scripts/check_framework_support_matrix.py` inside an activated venv with `uv sync --all-extras` already completed). A plain `python scripts/check_framework_support_matrix.py` from a fresh checkout will fail until dependencies are installed.
-**How**: The script imports each framework's test class, reflects its `supported_*()` set, and renders the tables. CI runs `--check` and fails if the on-disk matrix diverges from what the overrides now say.
+**Where**: The tables below are guarded by a pytest drift check in `mloda/community/feature_groups/data_operations/tests/test_framework_support_matrix.py`. Do not edit the block between the `BEGIN GENERATED` and `END GENERATED` markers by hand. After changing any `supported_*()` override, run `tox` (or at minimum that one test file). If the drift check fails, regenerate the block with a coding agent so its contents match what the test produces, then rerun until the test passes.
+**How**: The test reflects each framework's test class, reads its `supported_*()` set, renders the expected block, and asserts it matches the on-disk doc. CI runs `tox`, which runs the test, so a drifted matrix fails the build.
 
 ---
 
@@ -176,7 +176,7 @@ To promote a ✗ to ✓:
 1. Read the Known-divergences entry for the cell to understand why it is ✗.
 2. Implement the missing native expression (or remove the implementation-side `_validate_*_match` exclusion).
 3. Update the framework test class's `supported_*()` override, or remove it if the framework now matches the full default set.
-4. Run `python scripts/check_framework_support_matrix.py` so the generated block reflects the change.
+4. Rerun the pytest drift check (`pytest mloda/community/feature_groups/data_operations/tests/test_framework_support_matrix.py`). If it fails, regenerate the `BEGIN GENERATED` / `END GENERATED` block with a coding agent and rerun.
 5. Run `PYTEST_WORKERS=1 tox` to confirm the inherited tests now pass.
 
 ---
